refactor: rename and move block to be named "profile" instead of "gutenberg" * add render callback to block
